backend/requirements.txt
# Framework web
Flask==3.0.0
Flask-CORS==4.0.0

# Base de données
mysql-connector-python==8.2.0

# Sécurité
werkzeug==3.0.1
bcrypt==4.1.2

# Injection de dépendances
dependency-injector==4.41.0

# Tests
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utilitaires
python-dotenv==1.0.0
